refactor recipes tests in multiple modules
